chore: separate files with tasks
