+++ pkg/component/data/instillartifact/v0/task_sync_files.go
@@ -1,0 +1,7 @@
+package instillartifact
+
+import "google.golang.org/protobuf/types/known/structpb"
+
+func (e *execution) syncFiles(input *structpb.Struct) (*structpb.Struct, error) {
+	return nil, nil
+}
